+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - broken docs
+- lost items in `zip_sorted`
 
 ## [0.1.1] — 2025-10-13